account for non-present classes in label

+++ video_cv_project/data/common.py
@@ -45,7 +45,7 @@ def labels_to_tensor(ims: Sequence[Image.Image]):
     if pal:
         # cls: torch.Tensor = E.rearrange(lbls.unique(), "n -> n 1 1")
         # lbls = cls == E.repeat(lbls, "t h w -> t n h w", n=len(cls))
-        lbls = E.rearrange(one_hot(lbls), "t h w n -> t n h w")
+        lbls = E.rearrange(one_hot(lbls.long()), "t h w n -> t n h w")
         colors = E.rearrange(torch.tensor(pal), "(n c) -> n c", c=RGB)
 
     # `lbls` is THWC.
